-Fixed a bunch of warnings -Fixed a bug that was causing wrong VDP1 command information to be displayed
git-svn-id: https://yabause.svn.sourceforge.net/svnroot/yabause/trunk@1331 d7be59e3-070e-4b9a-85a1-4646befae34d

+++ yabause/src/windows/snddx.c
@@ -296,7 +296,7 @@ void SNDDXSetVolume(int volume)
    if (volume == 0)
       soundvolume = DSBVOLUME_MIN;
    else
-      soundvolume = log10((double)volume / 100.0) * 2000.0;
+      soundvolume = (LONG)(log10((double)volume / 100.0) * 2000.0);
 
    if (!issoundmuted)
       IDirectSoundBuffer8_SetVolume (lpDSB2, soundvolume);
